requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
markdown2==2.3.5
MarkupSafe==1.0
PyYAML==3.13
Werkzeug==0.14.1
